+++ app/controllers/recording_studio_stripe/usage_controller.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module RecordingStudioStripe
+  class UsageController < ApplicationController
+    before_action :authorize_view!
+
+    def show
+      @lines = billing.active_lines
+    end
+  end
+end